whack unnecessary cast

+++ src/intf-win32.c
@@ -186,7 +186,7 @@ _refresh_tables(intf_t *intf)
 	 * Map "unfriendly" win32 interface indices to ours.
 	 * XXX - like IP_ADAPTER_INFO ComboIndex
 	 */
-	for (i = 0; i < (int)intf->iftable->dwNumEntries; i++) {
+	for (i = 0; i < intf->iftable->dwNumEntries; i++) {
 		ifrow = &intf->iftable->table[i];
 		if (ifrow->dwType < MIB_IF_TYPE_MAX) {
 			_ifcombo_add(&intf->ifcombo[ifrow->dwType],
